Finished User layer logic
Finished User layer logic

+++ Internship_API/src/main/java/com/example/internship_api/service/BaseService.java
@@ -5,5 +5,5 @@ import com.example.internship_api.data.model.search_object.BaseSearchObject;
 
 public interface BaseService<TModel, TSearch extends BaseSearchObject> {
     PagedResult<TModel> getAll(TSearch search);
-    TModel getById(int id);
+    TModel getById(Long id);
 }
